Fixed various small issues found in testing

+++ futoin/cid/mixins/package.py
@@ -1,9 +1,12 @@
 
 from __future__ import print_function, absolute_import
 
-import os, tempfile
+import os, tempfile, subprocess, platform
 
 class PackageMixIn( object ):
+    def _isCentOS( self ):
+        return os.path.exists('/etc/centos-release')
+    
     def _isFedora( self ):
         return os.path.exists('/etc/fedora-release')
     
@@ -12,6 +15,12 @@ class PackageMixIn( object ):
 
     def _isArchLinux( self ):
         return os.path.exists('/etc/arch-release')
+    
+    def _isDebian( self ):
+        return platform.linux_distribution()[0] == 'debian'
+
+    def _isUbuntu( self ):
+        return platform.linux_distribution()[0] == 'Ubuntu'
     
     def _requireDeb(self, packages):
         apt_get = self._which('apt-get')
@@ -73,22 +82,37 @@ class PackageMixIn( object ):
             )
         
     def _addAptRepo(self, name, entry, gpg_key):
-        self._requireDeb(['software-properties-common', 'apt-transport-https', 'ca-certificates'])
+        self._requireDeb([
+            'software-properties-common',
+            'apt-transport-https',
+            'ca-certificates',
+            'lsb-release',
+        ])
         apt_add_repository = self._which('apt-add-repository')
         
         if not apt_add_repository:
             return
         
-        (fd, tf) = tempfile.mkstemp('cidgpg', text=True)
-        os.write(fd, gpg_key.encode(encoding='UTF-8'))
-        os.close(fd)
+        if gpg_key:
+            (fd, tf) = tempfile.mkstemp('cidgpg', text=True)
+            os.write(fd, gpg_key.encode(encoding='UTF-8'))
+            os.close(fd)
 
-        self._trySudoCall(
-            ['apt-key', 'add', tf],
-            errmsg = 'WARNING: you may need to import GPG key manually!'
-        )
+            self._trySudoCall(
+                ['apt-key', 'add', tf],
+                errmsg = 'WARNING: you may need to import GPG key manually!'
+            )
+            
+            os.remove(tf)
+            
+        codename = subprocess.check_output( ['lsb_release', '-cs'] )
         
-        os.remove(tf)
+        try:
+            codename = str(codename, 'utf8')
+        except:
+            pass
+        
+        entry = entry.replace('$codename$', codename)
             
         self._trySudoCall(
             [apt_add_repository, '--yes', entry],
@@ -101,6 +125,9 @@ class PackageMixIn( object ):
         )
             
     def _addRpmKey(self, gpg_key):
+        if not gpg_key:
+            return
+        
         rpm = self._which('rpm')
         
         if not rpm:
@@ -118,11 +145,11 @@ class PackageMixIn( object ):
         os.remove(tf)
             
     
-    def _addYumRepo(self, name, url, gpg_key):
+    def _addYumRepo(self, name, url, gpg_key=None):
         self._addRpmKey(gpg_key)
 
         dnf = self._which('dnf')
-        yumcfgmgr = self._which('yum-config-manager')
+        yum = self._which('yum')
         
         if dnf:
             self._requireYum(['dnf-plugins-core'])
@@ -130,13 +157,15 @@ class PackageMixIn( object ):
                 [dnf, 'config-manager', '--add-repo', url],
                 errmsg = 'WARNING: you may need to import GPG key manually!'
             )
-        elif yumcfgmgr:
+        elif yum:
+            self._requireYum(['yum-utils'])
+            yumcfgmgr = self._which('yum-config-manager')
             self._trySudoCall(
                 [yumcfgmgr, '--add-repo', url],
                 errmsg = 'WARNING: you may need to import GPG key manually!'
             )
 
-    def _addZypperRepo(self, name, url, gpg_key):
+    def _addZypperRepo(self, name, url, gpg_key=None):
         self._addRpmKey(gpg_key)
         
         zypper = self._which('zypper')
